S-81.0 step 4: LEVER A-BB6 — thread-local unit cache extended to the MAIN
design79 (emended S-80.0.7 + Council WP-82) implemented, every same-commit
obligation in THIS commit:

Core (php-runtime):
- main_unit_acquire/MainUnit/MainLever: SAME UnitKey cache (A-DS5, never a
  second cache); probe = canonicalize+stat (F3 revalidate_path semantics) +
  content fp; HIT returns (Rc<Module>, Rc<Program>), skips a1+a2.
- MAIN_CHAIN_FP COMPUTED (A-AH22/KS-AH-82-2): main_chain_fp() hashes the
  SAME PRELUDE_SRC binding lower_prelude consumes + all hoisted prelude
  units + reg_mode; in-cargo falsifier (mutated prelude => fp moves) PASS.
- fp domain: fp_mix(main_chain_fp, "main", source) — disjoint from include
  unit_fp; main entry carries main_program: Some(Rc<Program>), vacuously
  empty relocation (the main DEFINES the base id space).
- Publish: MainLever::publish_if_armed, ONE method both SAPIs, lexically
  AFTER link_fatal_check on clean link only (A-PP16; F8c counter form).
- park_main: the main's clone pins in the request RetainSet (design79 §4;
  A-DS8 gate expectation 1->2 BY NAME: lib+main).
- A-TH16/KH82-1: IC_EPOCH u32->u64 SAME COMMIT as the put (wrap now
  machine-impossible ~5.8e11 years; PropIc/MethodIc cells (u64,u32,u32,u32));
  DR-1 residual re-derived, gate PASS (3 gsub-counted tokens unchanged).
- A-DS14 pinned choices: (i) ini vector = EXECUTABLE empty-assert on the
  probed path; (ii) autoload-mid-compile = impossible for the main (lowers
  pre-Vm), F10 falsifies — no fp fold.
- A-DS15 declared at the pinned bump call-site (Vm::new only at request
  boundary; host-callback re-entry never constructs a fresh Vm).
- Counters (design79 par.8): main_probe/main_hit/main_put/main_impure_skip/
  main_probe_fail in UcStats + tag=unitcache row + uc_log closed vocabulary
  (union-build observable: fixture teeth are never census-only).
- A-TH14: probe is ONE parameter. Worker acquire(...,true); cli-server
  run_source_probed(...,true); one-shot delegates probe=false (corpus
  exercises the MISS-free path, F-oneshot pins main_probe==0).
- M-68.5 SUPERSEDED DELIBERATELY (A-DS11): include-put comment + Vm::module
  M-67.2 doc rewritten; old phrase BANNED in gate-doc-purge.
- A-BB22/KB-82-1: bare.php comment rewritten (a2(bare) is what the HIT
  REMOVES; bare-on-HIT is the measured ex-post floor); old phrase banned,
  doc-purge now scans fixtures/*.php too.

Gates (NEW, both PASS):
- wp81-harness/gate-lever-pins.sh: A-MS13/KS-MS-82-1 vm_new/RetainSet::new/
  park_main non-test allowlist per file + workspace sweep; A-PP16 order pin
  both SAPIs; KS-PP-82-3 SplitDrain-before-first-return; A-TH14 probed
  call-sites pinned; positive decoys for every detector.
- wp81-harness/gate-lever-fixtures.sh (A-PP17, armed same commit): F-oneshot
  THREE teeth (channel alive via reqmark on one-shot — new reqmark+flush on
  run_source_with_argv; main_probe==0; positive twin main_probe==nreq) +
  F5 (main_put==1, main_hit==nreq-1: the hit-counter MOVES) + F8b (link-fatal
  bodies byte-identical) + F8c COUNTER form (put==0/hit==0 across two
  link-fatals; fix => body ok, put==1). PASS on first full run.

design79 emended same commit: par.3 pinned choices + computed-fp text;
par.5 REWRITTEN in TWO stale classes (A-DS12) with per-field enumeration of
id-bearing payloads (ClassId/fn idx/pools/statics/slots — each
"deterministic from the virgin fp" or "Module-relative"); par.9 F13 row +
F4 deviation DECLARED (no impure-main path exists on this tree — Council
WP-83 judges) + F-oneshot three-teeth row.

Verified: workspace cargo test --release exit 0; axum tests 9/9 (A-DS8 at
2 BY NAME); census twin 17/17; php-runtime census 634/634; falsifier PASS;
capture-order, doc-purge (12 patterns), DR-1, lever-pins, lever-fixtures
ALL PASS. Corpus per NOME + full battery + matrix re-run = step 7 (A/B).

+++ php-rust/crates/php-runtime/src/bytecode.rs
@@ -155,7 +155,13 @@ thread_local! {
     /// requests, where the same numeric id can name a DIFFERENT class.
     /// Bumping the epoch at each `Vm` construction invalidates every cache
     /// in O(1). Starts at 1 and never returns 0 (0 = the empty-cache tag).
-    static IC_EPOCH: std::cell::Cell<u32> = const { std::cell::Cell::new(1) };
+    /// A-TH16/KH82-1 (Council WP-82, same commit as the main-unit put): u64 —
+    /// with the main Module cached process-lifetime its IC cells outlive
+    /// requests, so a u32 wrap (~4.3e9 bumps/thread, ~50 days at 1k req/s)
+    /// became an ACTIVE silent-stale class. At u64 the wrap needs ~5.8e11
+    /// years at the same rate: machine-impossible, residual closed. Cost:
+    /// 8 bytes per cell, zero atomics (thread-local).
+    static IC_EPOCH: std::cell::Cell<u64> = const { std::cell::Cell::new(1) };
 }
 
 /// Invalidate all [`PropIc`] caches (a new run's id space). Called by
@@ -168,7 +174,7 @@ pub fn bump_ic_epoch() {
 }
 
 #[inline]
-fn ic_epoch() -> u32 {
+fn ic_epoch() -> u64 {
     IC_EPOCH.with(|e| e.get())
 }
 
@@ -191,7 +197,7 @@ fn ic_epoch() -> u32 {
 /// due compilazioni identiche DEVONO confrontare uguali qualunque cosa la
 /// VM abbia cachato).
 #[derive(Debug)]
-pub struct PropIc(Rc<std::cell::Cell<(u32, u32, u32, u32)>>);
+pub struct PropIc(Rc<std::cell::Cell<(u64, u32, u32, u32)>>);
 
 impl PropIc {
     /// The cached `(class_id + 1, slot)` when filled IN THIS RUN for
@@ -247,7 +253,7 @@ impl Clone for PropIc {
 /// cloni dell'op, stato invisibile all'uguaglianza strutturale, epoch
 /// per-run contro gli id stantii.
 #[derive(Debug)]
-pub struct MethodIc(Rc<std::cell::Cell<(u32, u32, u32, u32)>>);
+pub struct MethodIc(Rc<std::cell::Cell<(u64, u32, u32, u32)>>);
 
 impl MethodIc {
     /// The cached `(defining ClassId, method idx)` when filled IN THIS RUN
